% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_checkpoint_heatmap.R
\name{gene_checkpoint_heatmap}
\alias{gene_checkpoint_heatmap}
\title{gene_checkpoint_heatmap}
\usage{
gene_checkpoint_heatmap(gene, method = "pearson")
}
\arguments{
\item{gene}{gene name likes "KLF7".}

\item{method}{method="pearson" is the default value. The alternatives to be passed to correlation are "spearman" and "kendall".}
}
\description{
Create a pan-cancer heatmap with symbols indicating statistical significance to reveal the correlation between the expression of a single gene and ICGs (immune checkpoint genes).
}
\examples{
gene_checkpoint_heatmap("KLF7")
}
