% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_deg_heatmap.R
\name{gene_deg_heatmap}
\alias{gene_deg_heatmap}
\title{gene_deg_heatmap}
\usage{
gene_deg_heatmap(cancer, gene, top_n = 20)
}
\arguments{
\item{cancer}{cancer name likes "BLCA".}

\item{gene}{gene name likes "KLF7".}

\item{top_n}{the number of top DEGS to be shown in the heatmap.}
}
\description{
Create a heatmap for differentially expressed genes grouped by the expression of a single gene in a specific type of cancer.
}
\examples{
gene_deg_heatmap("BLCA","KLF7")
}
