% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_gene_scatter.R
\name{gene_gene_scatter}
\alias{gene_gene_scatter}
\title{gene_gene_scatter}
\usage{
gene_gene_scatter(cancer, gene1, gene2, density = "F")
}
\arguments{
\item{cancer}{cancer name likes "BLCA".}

\item{gene1}{name of gene1 likes "CBX2".}

\item{gene2}{name of gene2 likes "CBX3".}

\item{density}{whether density of gene expression was shown.}
}
\description{
Scatter plot of gene and gene correlation in a specific type cancer.
}
\examples{
gene_gene_scatter("BLCA","CBX2","CBX3"),
          gene_gene_scatter("BLCA","CBX2","CBX3",density="T")
}
