% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_gsea_go.R
\name{gene_gsea_go}
\alias{gene_gsea_go}
\title{gene_gsea_go}
\usage{
gene_gsea_go(cancer, gene, logFC_cutoff = 2, pvalue_cutoff = 0.05)
}
\arguments{
\item{cancer}{cancer name likes "BLCA".}

\item{gene}{gene name likes "KLF7".}

\item{logFC_cutoff}{cutoff value of logFC, 2 was the default setting.}

\item{pvalue_cutoff}{cutoff value of pvalue, 0.05 was the default setting.}
}
\description{
GSEA-GO analysis of DEGs grouped by the expression of a single gene in a specific type of cancer, and the top 5 GO BP pathways were shown.
}
\examples{
gene_gsea_go("BLCA","KLF7")
}
