% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_immucell_heatmap.R
\name{gene_immucell_heatmap}
\alias{gene_immucell_heatmap}
\title{gene_immucell_heatmap}
\usage{
gene_immucell_heatmap(gene, method = "pearson")
}
\arguments{
\item{gene}{gene names likes "KLF7".}

\item{method}{method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall".}
}
\description{
Create a pan-cancer heatmap with symbols indicating statistical significance to reveal the correlation between the expression of a single gene and immnue cell ratio.
}
\examples{
gene_immucell_heatmap("KLF7")

}
