% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_immuscore_heatmap.R
\name{gene_immunescore_heatmap}
\alias{gene_immunescore_heatmap}
\title{gene_immunescore_heatmap}
\usage{
gene_immunescore_heatmap(gene, method = "pearson")
}
\arguments{
\item{gene}{gene names likes "KLF7".}

\item{method}{method="pearson" is the default value. The alternatives to be passed to correlation are "spearman" and "kendall".}
}
\description{
Create a pan-cancer heatmap with symbols indicating statistical significance to reveal the correlation between the expression of a single gene and immune scores, including Stromal score, immune score, and ESTIMATE score.
}
\examples{
gene_immunescore_heatmap("KLF7")

}
