% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_immuscore_triangle.R
\name{gene_immunescore_triangle}
\alias{gene_immunescore_triangle}
\title{gene_immunescore_triangle}
\usage{
gene_immunescore_triangle(gene, method = "pearson")
}
\arguments{
\item{gene}{gene name likes "KLF7".}

\item{method}{method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall".}
}
\description{
Create a pan-cancer triangle reveals the correlation between the expression of a single gene and immune scores, including Stromal score, immune score, and ESTIMATE score.
}
\examples{
gene_immunescore_triangle("KLF7")
}
