% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_immustimulator_heatmap.R
\name{gene_immustimulator_heatmap}
\alias{gene_immustimulator_heatmap}
\title{gene_immustimulator_heatmap}
\usage{
gene_immustimulator_heatmap(gene, method = "pearson")
}
\arguments{
\item{gene}{gene names likes "KLF7".}

\item{method}{method="pearson" is the default value. The alternatives to be passed to correlation are "spearman" and "kendall".}
}
\description{
Create a pan-cancer heatmap with symbols indicating statistical significance to reveal the correlation between the expression of a single gene and immune stimulators.
}
\examples{
gene_immustimulator_heatmap("KLF7")
}
