% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_methylation_scatter.R
\name{gene_methylation_scatter}
\alias{gene_methylation_scatter}
\title{gene_methylation_scatter}
\usage{
gene_methylation_scatter(cancer, gene)
}
\arguments{
\item{cancer}{cancer name likes "BLCA".}

\item{gene}{gene name likes "KLF7".}
}
\description{
Scatter plot of gene expression and gene promoter methylation correlation in a specific type of cancer. A pdf file named gene_methylation will be generated in the working directory.
}
\examples{
gene_methylation_scatter("BLCA","KLF7")
}
