% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_MSI_radar.R
\name{gene_MSI_radar}
\alias{gene_MSI_radar}
\title{gene_MSI_radar}
\usage{
gene_MSI_radar(gene, method = "pearson")
}
\arguments{
\item{gene}{gene name}

\item{method}{method="pearson" is the default value. The alternatives to be passed to correlation are "spearman" and "kendall".}
}
\description{
Create a pan-cancer radar chart for gene expression and MSI correlation.
}
\examples{
gene_MSI_radar("KLF7")
}
