% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_TMB_radar.R
\name{gene_TMB_radar}
\alias{gene_TMB_radar}
\title{gene_TMB_radar}
\usage{
gene_TMB_radar(gene, method = "pearson")
}
\arguments{
\item{gene}{gene name likes "KLF7".}

\item{method}{method="pearson" is the default value. The alternatives to be passed to correlation are "spearman" and "kendall".}
}
\description{
Create a pan-cancer radar chart for gene expression and TMB correlation.
}
\examples{
gene_TMB_radar("KLF7")
}
