% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{get_meta}
\usage{
get_meta(cancer)
}
\arguments{
\item{cancer}{cancer name likes "COAD".}
}
\description{
Extract the clinical information of a specific type of cancer in TCGA.
}
\examples{
get_meta("COAD")
}
