% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paired_cancers.R
\name{get_paired_cancers}
\alias{get_paired_cancers}
\title{get_paired_cancers}
\usage{
get_paired_cancers()
}
\description{
Return the sample summary of 15 types of cancer containing more than 20 paired samples in TCGA.
}
\examples{
get_paired_cancers()
}
