% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paired_tpm.R
\name{get_paired_tpm}
\alias{get_paired_tpm}
\title{get_paired_tpm}
\usage{
get_paired_tpm(cancer)
}
\arguments{
\item{cancer}{cancer name likes "COAD".}
}
\description{
Extract the TPM matrix of a specific type of cancer with paired samples (n>20) in TCGA.
}
\examples{
get_paired_tpm("COAD")
}
