% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tpm.R
\name{get_tpm}
\alias{get_tpm}
\title{get_tpm}
\usage{
get_tpm(cancer)
}
\arguments{
\item{cancer}{cancer name likes "COAD".}
}
\description{
Extract the TPM matrix of a specific type of cancer in TCGA.
}
\examples{
get_tpm("COAD")
}
