% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_boxplot.R
\name{paired_boxplot}
\alias{paired_boxplot}
\title{paired_boxplot}
\usage{
paired_boxplot(cancer, gene, palette = "jco", legend = "none")
}
\arguments{
\item{cancer}{name of cancer likes "BRCA"}

\item{gene}{name of gene like "KLF7"}

\item{palette}{the color palette to be used for coloring or filling by groups. Allowed values include scientific journal palettes from ggsci R package, e.g.: "npg", "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and "rickandmorty".}

\item{legend}{legend position. Allowed values include "top","bottom","left","right" and "none".}
}
\description{
Create a paired tumor-normal box plot for a single gene with symbols indicating statistical significance in a specific type of cancer. Only cancers with more than 20 paired samples could be analyzed, including "BLCA","BRCA","COAD","ESCA","HNSC","KICH","KIRC","KIRP","LIHC","LUAD","LUSC","PRAD","STAD","THCA", and "UCEC".
}
\examples{
paired_boxplot("BRCA","KLF7")
}
