% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pan_boxplot.R
\name{pan_boxplot}
\alias{pan_boxplot}
\title{pan_boxplot}
\usage{
pan_boxplot(gene, palette = "jco", legend = "right")
}
\arguments{
\item{gene}{gene names likes "KLF7".}

\item{palette}{the color palette to be used for filling by groups. Allowed values include scientific journal palettes from ggsci R package, e.g.: "npg", "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and "rickandmorty".}

\item{legend}{legend position. Allowed values include "top","bottom","left","right" and "none".}
}
\description{
Create a pan-cancer box plot for a single gene with symbols indicating statistical significance.
}
\examples{
pan_boxplot("KLF7")
}
