% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pan_forest.R
\name{pan_forest}
\alias{pan_forest}
\title{pan_forest}
\usage{
pan_forest(gene)
}
\arguments{
\item{gene}{gene name likes "KLF7".}
}
\description{
Create a pan-cancer Cox regression forest plot for a specific gene.
}
\examples{
pan_forest("KLF7")
}
