% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pan_paired_boxplot.R
\name{pan_paired_boxplot}
\alias{pan_paired_boxplot}
\title{pan_paired_boxplot}
\usage{
pan_paired_boxplot(gene, palette = "jco", legend = "right")
}
\arguments{
\item{gene}{gene name likes "KLF7"}

\item{palette}{the color palette to be used for filling by groups. Allowed values include scientific journal palettes from ggsci R package, e.g.: "npg", "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and "rickandmorty".}

\item{legend}{legend position. Allowed values include "top","bottom","left","right" and "none".}
}
\description{
Create a pan-cancer paired box plot for a single gene with symbols indicating statistical significance.
}
\examples{
pan_paired_boxplot("KLF7")
}
