% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_boxplot.R
\name{tcga_boxplot}
\alias{tcga_boxplot}
\title{tcga_boxplot}
\usage{
tcga_boxplot(cancer, gene, add = "jitter", palette = "jco", legend = "none")
}
\arguments{
\item{cancer}{cancer name likes "BRCA".}

\item{gene}{gene name likes "KLF7".}

\item{add}{character vector for adding another plot element likes "none", "dotplot", "jitter".}

\item{palette}{the color palette to be used for coloring or filling by groups. Allowed values include scientific journal palettes from ggsci R package, e.g.: "npg", "aaas", "lancet", "jco".}

\item{legend}{legend position. Allowed values include "top","bottom","left","right" and "none".}
}
\description{
Create a tumor-normal box plot for a single gene with symbols indicating statistical significance in a specific type of cancer.
}
\examples{
tcga_boxplot("BRCA","KLF7")
}
