% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_kmplot.R
\name{tcga_kmplot}
\alias{tcga_kmplot}
\title{tcga_kmplot}
\usage{
tcga_kmplot(cancer, gene, palette = "jco")
}
\arguments{
\item{cancer}{cancer name likes "COAD".}

\item{gene}{gene name likes "KLF7".}

\item{palette}{the color palette to be used for coloring or filling by groups. Allowed values include scientific journal palettes from ggsci R package, e.g.: "npg", "aaas", "lancet", "jco".}
}
\description{
K_M survival plot for a single gene in a specific type of cancer.
}
\examples{
tcga_kmplot("COAD","KLF7")
}
