% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_roc.R
\name{tcga_roc}
\alias{tcga_roc}
\title{tcga_roc}
\usage{
tcga_roc(cancer, gene)
}
\arguments{
\item{cancer}{cancer name likes "BRCA".}

\item{gene}{gene name likes "KLF7".}
}
\description{
Diagnostic ROC curve of a single gene in a specific type of cancer.
}
\examples{
tcga_roc("BRCA","KLF7")
}
